+++ src/components/Toolbar.js
@@ -1,0 +1,24 @@
+import React from "react";
+
+function Toolbar() {
+  return (
+    <div className="border-b-2 flex flex-row">
+      <button type="button" className="menu">
+        File
+      </button>
+      <button type="button" className="menu">
+        Edit
+      </button>
+      <button type="button" className="menu">
+        Inspect
+      </button>
+      <button type="button" className="menu">
+        View
+      </button>
+      <button type="button" className="menu">
+        Share
+      </button>
+      <div />
+    </div>
+  );
+}
