+++ src/pages/main.js
@@ -5,8 +5,12 @@ import Header from "../components/header";
 export default function Main() {
   return (
     <div className="h-screen flex flex-col">
-      <Header />
-      <Composer />
+      <div className="flex-shrink-0 flex-grow-0">
+        <Header />
+      </div>
+      <div className="flex-grow">
+        <Composer />
+      </div>
     </div>
   );
 }
